Add --init-alloc-size parameter for fast initial allocation
Often, it is useful to first allocate a big chunk of memory
before continuing to gradually allocate smaller chunks
to cuase stress.

For example, let's say that someone is trying to hit a
certain memory limit of 5GB. In such a case, it is often
useful to first allocate 4GBs in one big chunk fastly,
and then continue stressing the system using allocation
size of 16MB each 5 seconds.

After this PR, the following is possible:
stress -mem-total 5Gi -mem-alloc-size 16Mi -mem-alloc-sleep 5s
-init-alloc-size 4Gi

Signed-off-by: Itamar Holder <iholder@redhat.com>